+++ home-manager/default.nix
@@ -1,6 +1,7 @@
 {
   imports = [
     ./modules/git.nix
+    ./modules/gpg.nix
     ./modules/zsh.nix
   ];
 }